Fix database type error and missing await in verification
1. Fix credibility_score column type:
   - Change from INTEGER to NUMERIC(3,1) to accept decimal values
   - Update schema and create migration SQL
   - Fixes 'invalid input syntax for type integer' errors

2. Add missing await calls in verification module:
   - await artistOps.findById()
   - await sourceOps.findByArtistId()
   - await artistOps.updateStatus()
   - await artistOps.findByStatus()
   - Fixes 'discovered is not iterable' error

All database operations now properly awaited.

+++ supabase-schema.sql
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS sources (
   artist_id INTEGER NOT NULL REFERENCES artists(id) ON DELETE CASCADE,
   url TEXT NOT NULL,
   institution TEXT,
-  credibility_score INTEGER DEFAULT 5,
+  credibility_score NUMERIC(3,1) DEFAULT 5.0,
   content_summary TEXT,
   created_at TIMESTAMP DEFAULT NOW()
 );
